Add CRUD-operations, func info

+++ src/primitive_db/engine.py
@@ -2,26 +2,10 @@ import shlex
 
 import prompt
 
-from .color_text import bold_text, red_italics_text, underlined_text
+from .color_text import red_italics_text
 from .constants import COMMANDS, INPUT_PROMT, META_FILE
 from .core import Database
-from .utils import load_metadata
-
-
-def print_help(commands: dict) -> None:
-    """
-    Выводит список доступных функций и их описания.
-
-    Args:
-        commands (dict): словарь функций и их описаний
-    """
-
-    print(bold_text("\n***   Процесс работы с таблицей   ***"))
-    print(underlined_text("Функции:"))
-    for i, v in commands.items():
-        print(f"{i:<25} — {v}")
-    print()
-
+from .utils import load_metadata, print_help
 
 
 def process_command(db: Database, metadata: dict, cmd: str):
@@ -34,21 +18,56 @@ def process_command(db: Database, metadata: dict, cmd: str):
         cmd (str): Введенная пользователем строка с командой и аргументами.
     """
 
-    split_cmd = shlex.split(cmd)
-    match split_cmd[0]:
+    split_cmd = shlex.split(cmd.strip())
+
+    match split_cmd[0].lower():
         case "create_table":
+            if len(split_cmd) < 3:
+                raise ValueError(f"Некорректная функция: {cmd}. Попробуйте снова.")
             columns = [col.split(":") for col in split_cmd[2:]]
             db.create_table(metadata, split_cmd[1], columns)
+        
         case "list_tables":
             db.list_tables(metadata)
+        
         case "drop_table":
+            if len(split_cmd) < 2:
+                raise ValueError(f"Некорректная функция: {cmd}. Попробуйте снова.")
             db.drop_table(metadata, split_cmd[1])
+        
+        case "insert":
+            if len(split_cmd) < 5:
+                raise ValueError(f"Некорректная функция: {cmd}. Попробуйте снова.")
+            db.insert(metadata, split_cmd[2], split_cmd[4:])
+
+        case "select":
+            if len(split_cmd) < 3:
+                raise ValueError(f"Некорректная функция: {cmd}. Попробуйте снова.")
+            db.select(metadata, split_cmd[2], split_cmd[3:])
+
+        case "update":
+            if len(split_cmd) < 6:
+                raise ValueError(f"Некорректная функция: {cmd}. Попробуйте снова.")
+            db.update(metadata, split_cmd[1], split_cmd[2:])
+
+        case "delete":
+            if len(split_cmd) < 4:
+                raise ValueError(f"Некорректная функция: {cmd}. Попробуйте снова.")
+            db.delete(metadata, split_cmd[2], split_cmd[3:])
+
+        case "info":
+            if len(split_cmd) < 2:
+                raise ValueError(f"Некорректная функция: {cmd}. Попробуйте снова.")
+            db.info(metadata, split_cmd[1])
+
         case "help":
             print_help(COMMANDS)
+
         case "exit":
             exit()
+
         case _:
-            print(red_italics_text(f"Функции <{cmd}> нет. Попробуйте снова.\n"))
+            raise ValueError(f"Функции <{split_cmd[0]}> нет. Попробуйте снова.")
 
 
 def run():
@@ -60,6 +79,11 @@ def run():
 
     db = Database()
     while True:
-        metadata = load_metadata(META_FILE)
-        user_input = prompt.string(INPUT_PROMT).strip()
-        process_command(db, metadata, user_input)
+        try:
+            metadata = load_metadata(META_FILE)
+            user_input = prompt.string(INPUT_PROMT)
+            process_command(db, metadata, user_input)
+        except (ValueError) as error:
+            print(red_italics_text(error))
+        except (KeyboardInterrupt, EOFError):
+            break
